+++ cmd/build.go
@@ -145,7 +145,7 @@ func runBuild(cmd *cobra.Command, _ []string, clientFn buildClientFn) (err error
 	}
 
 	// All set, let's write changes in the config to the disk
-	err = function.WriteConfig()
+	err = function.Write()
 	if err != nil {
 		return
 	}
