+++ schema/func_yaml-schema.json
@@ -1,25 +1,42 @@
 {
 	"$schema": "http://json-schema.org/draft-04/schema#",
-	"$ref": "#/definitions/Config",
+	"$ref": "#/definitions/Function",
 	"definitions": {
-		"Config": {
+		"Env": {
+			"required": [
+				"value"
+			],
+			"properties": {
+				"name": {
+					"pattern": "^[-._a-zA-Z][-._a-zA-Z0-9]*$",
+					"type": "string"
+				},
+				"value": {
+					"type": "string"
+				}
+			},
+			"additionalProperties": false,
+			"type": "object"
+		},
+		"Function": {
 			"required": [
 				"name",
 				"namespace",
 				"runtime",
+				"Registry",
 				"image",
 				"imageDigest",
 				"builder",
 				"builders",
 				"buildpacks",
-				"healthEndpoints",
 				"volumes",
 				"buildEnvs",
 				"envs",
 				"annotations",
 				"options",
 				"labels",
-				"created"
+				"healthEndpoints",
+				"Created"
 			],
 			"properties": {
 				"name": {
@@ -30,6 +47,9 @@
 					"type": "string"
 				},
 				"runtime": {
+					"type": "string"
+				},
+				"Registry": {
 					"type": "string"
 				},
 				"image": {
@@ -54,10 +74,6 @@
 						"type": "string"
 					},
 					"type": "array"
-				},
-				"healthEndpoints": {
-					"$schema": "http://json-schema.org/draft-04/schema#",
-					"$ref": "#/definitions/HealthEndpoints"
 				},
 				"volumes": {
 					"items": {
@@ -98,25 +114,13 @@
 					},
 					"type": "array"
 				},
-				"created": {
+				"healthEndpoints": {
+					"$schema": "http://json-schema.org/draft-04/schema#",
+					"$ref": "#/definitions/HealthEndpoints"
+				},
+				"Created": {
 					"type": "string",
 					"format": "date-time"
-				}
-			},
-			"additionalProperties": false,
-			"type": "object"
-		},
-		"Env": {
-			"required": [
-				"value"
-			],
-			"properties": {
-				"name": {
-					"pattern": "^[-._a-zA-Z][-._a-zA-Z0-9]*$",
-					"type": "string"
-				},
-				"value": {
-					"type": "string"
 				}
 			},
 			"additionalProperties": false,
